Refactor thermal_control_test_helper.py
