+++ tests/platform_tests/api/test_chassis_fans.py
@@ -5,7 +5,7 @@ import pytest
 
 from tests.common.helpers.platform_api import chassis, fan
 from .platform_api_test_base import PlatformApiTestBase
-from tests.platform_tests.thermal_control_test_helper import start_thermal_control_daemon, stop_thermal_control_daemon
+from tests.common.helpers.thermal_control_test_helper import start_thermal_control_daemon, stop_thermal_control_daemon
 
 ###################################################
 # TODO: Remove this after we transition to Python 3
