+++ tests/platform_tests/test_platform_info.py
@@ -14,9 +14,8 @@ from tests.common.helpers.assertions import pytest_assert, pytest_require
 from tests.common.plugins.loganalyzer.loganalyzer import LogAnalyzer
 from tests.common.utilities import wait_until, get_sup_node_or_random_node
 from tests.common.platform.device_utils import get_dut_psu_line_pattern
-from .thermal_control_test_helper import ThermalPolicyFileContext,\
-    check_cli_output_with_mocker, restart_thermal_control_daemon, check_thermal_algorithm_status,\
-    mocker_factory, disable_thermal_policy  # noqa F401
+from tests.common.helpers.thermal_control_test_helper import ThermalPolicyFileContext,\
+    check_cli_output_with_mocker, restart_thermal_control_daemon, check_thermal_algorithm_status  # noqa F401
 
 pytestmark = [
     pytest.mark.topology('any')
